backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
aiohttp==3.8.4
openai==0.27.8
pillow==9.5.0
numpy==1.24.3
transformers==4.30.2
torch==2.0.1
torchvision==0.15.2
python-dotenv==1.0.0
pydantic==1.10.8
httpx==0.24.1 